added placeholders for build id and revision in about text
They will be replaced in the build process.

Attention: The value for the about text in the product file
is not used during the build. The information is stored in
the product definition in the plugin.xml in the application
bundle. When updating the about text using the product editor
you have to run the product in Eclipse for it to update the
plugin.xml file.

Change-Id: I0b8b1dfea7dfa36560048c4c809c9a8ad287d8f7
Reviewed-on: http://gerrit.esdi-community.eu/210
Tested-by: Jenkins <sitemple@igd.fraunhofer.de>
Reviewed-by: Simon Templer <simon.templer@igd.fraunhofer.de>