added clearCache method to PropertyResolver
This is just a short term fix to the problem that the
PropertyResolver cache may be outdated. PropertyResolver
and/or its cache should be available as a service and
reset on schema changes. It might be a good idea allowing
instantiation of a local property resolver w/ cache.
Also there might be problems with differences in source
and target types that are equally named but can't be
differentiated by PropertyResolver.

Change-Id: I6dcecf9380c253728c874ec9d66636583ff0538c