added method to get a filtered selection to InstanceCollection
In InstanceCollection implementations a very simple solution
is used for now, where a FilteredInstanceCollection wraps the
InstanceCollection an applies the filter to each Instance element.
This could be further optimized, especially where the OrientDB
or GML based InstanceCollections are concerned (e.g. adapting a
DB query or skipping GML instances if the type does not match).

refs #312

Change-Id: I43c3b1a8b3d0325927e8512b17679451738408dd