XML reader: added support for non-strict property matching
This allows to load files that don't strictly follow the
schema structure, though the representation in the model
may then also be not according to the schema.
This is not an optimal solution - it might be good to check
if this can be improved, e.g. by still checking the cardinality
but not the order, and thus providing instances that are
compatible to the schema as much as possible.

refs #304

Change-Id: I4a7fd9ff9b996c66d42b275445cf46635c6d2347
Reviewed-on: http://gerrit.esdi-community.eu/824
Tested-by: Jenkins <sitemple@igd.fraunhofer.de>
Reviewed-by: Simon Templer <simon.templer@igd.fraunhofer.de>