use a login page instead of basic auth by default
Using the login form now a login to all wars is performed when
logging in to one war (those that use Secured***ApplicationContext).
This is achieved by using a remember-me cookie, with / as fixed
path for all wars. The cookie only lives while the browser runs,
also the remember me key token is generated randomly for the life
of the JVM instance - so there is no danger of different HALE
server instances being authenticated through the same cookie.
Logout is performed using a special service, that tracks sessions
associated to a user in different wars, and invalidates all of them
during logout.

refs #507

Change-Id: I24151023e411968f0284dfe8b6733de077d15120