use shorter class names for OrientDB storage for groups
Before for groups the identifier was used as class name, which
could be very long at times, which caused errors creating database
files because the file name was too long.
Now the group qualified name will be used, of which the namespace
is recplaced with a short prefix when encoded.

Change-Id: I1835dd2de56dbc5208c017f78f411a5a4b57e193