about/contact - replaced unicode characters w/ HTML entities
Because for some unknown reason when behind the proxy on
hale.igd.fraunhofer.de, browsers don't display the pages correctly.
The encoding is communicated correctly both in the HTTP header and
in the document, so this probably is an issue with the proxy
configuration - but at the moment easier to solve this way.

Change-Id: Ic05fb042432f0ef9f585a3bb504a5098c2e976c3
Reviewed-on: http://gerrit.esdi-community.eu/2454
Tested-by: Jenkins <sitemple@igd.fraunhofer.de>
Reviewed-by: Simon Templer <simon.templer@igd.fraunhofer.de>