target platform update - added Jetty 9
The patched Jetty 9 is needed for the server applications
to be able to run.
This introduces two Jetty versions to the target platform.
The products for HALE and Infocenter are configured to use
Jetty 8 as this is needed for the help. However, the tooling
in Eclipse does not correctly support creating the run
configuration based on the product, selecting the right versions.
Thus, when creating the Run configuration from the product,
it is necessary to fix it manually.

Change-Id: I93dd911f918b221c6231928d5c62a93b105aa8f1
Reviewed-on: http://gerrit.esdi-community.eu/2926
Tested-by: Jenkins <sitemple@igd.fraunhofer.de>
Reviewed-by: Simon Templer <simon.templer@igd.fraunhofer.de>