Added support to read XML file and parse XML out of XML file.
To declare an XML Parameter Use -S (Setting Prefix) if you need a path to
an XML file. If -X is used the execution tries to parse the content of the
xml file (the path is given as a parameter to the application), so
provider parameter such as inspire.sds.metadata.inline of
InspireInstanceWriter.

There are 3 things to mention at this state:
1. The targetFile differs between metadata and metadata.inline parameter
usage.
2. If the XML file contains any line break, empty lines will be added
after each metadata line in the targetfile. But the metadata is correct.
3. The path of the parameter used for testing... @see ExecuteTest

Change-Id: I8cce8026c084660db27b9754cefc1b12cb47887e