added first version of CleanupService
That performs application or project scoped cleanup tasks.
Either with a custom implementation of Cleanup or removing given temporary
files. The temporary files list is persisted so the files can be removed
even after the program has crashed, and is started again.

Open issue: This will only work like this for a singe running instance of
HALE, but will lead to strange results if multiple instances are running.

Change-Id: If44f054aeba524526db06b06e0bef5042c833eaf
Reviewed-on: http://gerrit.esdi-community.eu/3112
Reviewed-by: Simon Templer <simon.templer@igd.fraunhofer.de>
Tested-by: Simon Templer <simon.templer@igd.fraunhofer.de>