build: automatic RELEASE qualifier only for matching application bundles
Applying the RELEASE qualifier to all bundles/features would require to
consistently increase the version number for every bundle. But there are
currently bundles where the version number is not increased for each
release, if they specify a qualifier in the version using a fixed
qualifier would break the versioning.

Change-Id: Icad6e5dc263ab6da8a68a47736ff42d93bca7132
Reviewed-on: http://gerrit.esdi-community.eu/3173
Tested-by: Jenkins <sitemple@igd.fraunhofer.de>
Reviewed-by: Simon Templer <simon.templer@igd.fraunhofer.de>