Fix resource copying in archive writer
The way the resources were copied to the temporary directory
before could cause problems with files that contained
character sequences that can be interpreted as being URL-
encoded (like `%20`). In that case, the file name of the copied
resource did not match the relative resource path in the
save configuration.