Fix updating parentType of property selectors after type change
When the source or target type of a cell is changed, the parentType of the
corresponding property selectors needs to be updated to allow subsequent
selection of the properties of the newly selected type.

This update failed because the method used to update the parentType has side
effects, i.e. it modifies the list of property selectors to remove previously
selected properties when the parentType changes.

This fix replaces the iterator-based approach of updating the property
selectors by one that is safe with regards to the side effects of
setParentType.