Make Request class usable in headless mode
In headless mode the Request class could not be used up to now
to download resources b/c initialization failed in non-OSGi environments..
This poses a problem in cases where the fallback mechanism applied
by the caller (if any) is not sophisticated enough to handle e.g. proxies
or HTTP redirects (e.g. URL.openStream()).