Improve CRS detection when parsing GML
When an `srsName` attribute is encountered in the GML stream, the
corresponding CRSDefinition will be resolved and passed down the
element hierarchy to be used by nested geometries that don't have
their own `srsName` attribute.

Additionally, if an `srsName` cannot be resolved, the encountered
code will now be passed to the `CRSResolver` and, in case the GUI is
present, be displayed in the CRS selection dialog.