Allow importing Shapefiles with shortened property names
Match property names in Shapefile to target type properties by checking
if there is exactly one target type property name that starts with the
Shapefile property name. If there are multiple property names in the
target type that start with the Shapefile property name, the property
will not be imported.