perf(cst): use thread-local Groovy script cache
The method of storing compiled scripts in the `ExecutionContext` of
either the `TransformationFunction` or the `Cell` does not work in the
case of inline transformations. If a property is mapped by an inline
transformation, a separate transformer instance is instantiated for
every property value. As the `ConceptualSchemaTransformer` itself is
stateless, the cached scripts do not get passed on and are recompiled
for every property value, causing a significant slowdown.

This new simple thread-local cache avoids this problem as long as the
CSTs for the inline property transformations are all executed within
the same thread.

A possible drawback is the increased memory usage as, over the course of
a transformation run, every Groovy script will be cached. If this turns
out to be a problem, it can be mitigated by using a more sophisticated
cache implementantion with e.g. an upper memory limit.

https://github.com/halestudio/hale/issues/575