fix(io): prevent bad IOConfiguration paths after hale connect export
When exporting a project to hale connect, the resource URIs in the
`IOConfiguration` would point to the temporary location where the
project archive for the hc upload was assembled. These temporary paths
would then be persisted if the project was subsequently saved locally.
It would also lose all source data configurations if the hc upload
did not contain source data.

This patch prevents this by restoring the `IOConfiguration`s if the
project is exported to hale connect (i.e. `HaleConnectProjectWriter`
is run in `EXPORT` mode).

https://github.com/halestudio/hale/issues/642