feat(cst): support joins of float and integer properties
This patch adds the same behaviour to the processing of double and
float values that already existed for BigDecimals. In particular,
double and float values that are in fact integers (e.g. `2.0`) are
now serialized without the trailing `.0` during value processing.

This allows the mixed use of float/double and integer/string properties
in Joins.

https://github.com/halestudio/hale/issues/737