fix(hc): handle cases where user/org information is unavailable
This adds handling for cases where user or organisation info is
queried via the hale connect API, but the information is cannot
be obtained (e.g. because the user/org was deleted in the meantime
or because of missing authorization).

Previously, failure to retrieve user/org information could lead
to errors in the UI.

https://github.com/halestudio/hale/issues/752