feat: support creating GeoPackage files and writing instances
Basic support for writing instances to a new or existing GeoPackage
file.
If the writer encounters an existing table it will write to that table
(and not change its structure). If the writer does not find a table it
tries to create it. The writer supports feature and attribute tables.

In addition added support for attribute tables to GeoPackage schema reader and
instance reader (to complement the writer functionality).

Possible later improvements to this basic version:

- specific support for Date and DateTime data types
- support foreign key constraints
- automatically fill foreign keys from autogenerated primary keys
  (probably a prerequisite for most scenarios where foreign keys are
  used)
- support database transactions
- support geometry types that are not supported by the JTS library used
  in hale