fix: allow filtering LimboInstanceSink instance collection
Not sure why this was not implemented before - it should be OK just
retrieving specific instances, even if it is not possible to iterate
over the collection more than once. This just means that there might be
instances coming out of the transformation that are not
handled/exported.

The previous behavior would prevent instance to be written with a CSV or
Excel writer in case the LimboInstanceSink was used, as it only processes
instances of a specific type (depending on configuration).