fix(io): List binding: trim whitespace before splitting
Make sure that leading and trailing whitespace is removed before splitting
the value of an element bound to `java.util.List` at whitespace.

The old behaviour lead to three list entries (an empty one followed by the one
for each of the actual values) being created for the `pos` element value from
the following snippet:

```
	<gml:Point srsName="EPSG:25832" gml:id="GML-f">
		<gml:pos>
			540465.8278 5802259.4572
		</gml:pos>
	</gml:Point>
```

However, the snippet above should have the same semantics as

```
	<gml:Point srsName="EPSG:25832" gml:id="GML-f">
		<gml:pos>540465.8278 5802259.4572</gml:pos>
	</gml:Point>
```

Therefore, leading and trailing whitespace is now removed before the split.

https://github.com/halestudio/hale/issues/808