refactor(geopackage): use INTEGER data type for BigInteger binding
Because the BigInteger binding is used in XML Schemas in hale as binding
for any integer, interpret it as integer here as well.
This means that XML Schemas can be more easily used as target schemas
when writing a Geopackage, but it also means that there will be
information lost if there are actually BigIntegers outside the range of
long.

Also added a general conversion to the Geopackage column type to avoid
problems when saving a row.

refs #788