feat(geopackage): add support for selecting a target CRS
A target CRS can now be specified as a parameter. It will be used as a
reference system for geometries for newly created tables. If the schema
has SRS information associated, that information will be preferred over
the provide CRS. Fall-back stays the CRS found in the first instance
encountered for a type.

The key issue was that by default there do not seem to be many CRS
stored within the Geopackage. So except for a few (like WGS84) they need
to be created first. I hope this is done how it should be - at least my
tests in QGIS were successful, displaying a Geopackage where the CRS was
inserted from the Geotools definition.

Note: Currently only CRS w/ EPSG code are supported to be added to the
Geopackage.

refs #789