fix: use correct preferred bundle when loading geometry types
As part of the upgrade to GeoTools 21, a wrong bundle name was used as the
preferred bundle when loading geometry types. The wrong bundle name causes
significant delays when loading schemas with geometries as hale tried to
locate the non-existing bundle for every geometry element in the schema.

The correct bundle name is `org.locationtech.jts.jts-core`. In addition,
replacing the old GeoTools package name in the class name is now done
before trying to load geometry class.

https://github.com/halestudio/hale/issues/861