feat(gpkg): allow to create spatial index when writing GeoPackage
Adds a configuration option to the GeoPackage writer that allows to define
which type of spatial index (if any) should be created when feature tables
are newly created in the target GeoPackage file.

Also changes the default behaviour of the GeoPackage writer so that an
RTree spatial index is now created by default if no other option is
specified.

https://github.com/halestudio/hale/issues/883