fix: fix ArcString interpolation handling
When an ArcString object was devided in single Arc objects in order to
interpolate them before, the iterator was raised by 3 in every step. As
gml:ArcString objects consist of multiple arcs object and the endpoint
of one arc represents the start point of the next arc at the same time.
This means, if two arcs are contained in an ArcString object, the first
arc starts as coordinate 0 and the second one starts at coordinate 2 etc.
So the iterator for the starting point should only be raised by 2 in order
to access the next arc.

WGS-1215