fix: prevent multiple imports of the same schema
As nowadays the two URIs "http://example.org/a.xsd" and
"https://example.org/a.xsd" typically refer to the same resource (the former
being redirected to the latter), add awareness for this to hale studio's XML
schema reader to prevent multiple imports of the same schema.

Without this patch, the following situation can occur:

Let A be an XML schema available both via "http://example.org/a.xsd" and
"https://example.org/a.xsd". A imports another XML schema B. B, in turn,
imports A again, using the schema location "https://example.org/a.xsd".

If A is now loaded into hale studio from "http://example.org/a.xsd", the XML
schema reader will parse A, load and parse B, and then load and parse A again
because the schema location of the original import ("http://example.org/a.xsd")
differs from the schema location used in the import tag in B
("https://example.org/a.xsd"). This result in duplicate schema elements in
hale's schema explorer.