fix: fix validity check of the hale»connect export dialog
Because permissions checks done as part of the login status update are executed
asynchronously, it is possible that the owner radio buttons are only activated
after the validity of the dialog is input is already determined. This leads to
the situation that a logged-in user can't create or update transformation
projects even if they have the correct permissions.

With this fix, the login status check and dialog validation are separated and
the dialog validation is retriggered once the asynchronous operation to
determine the user's permissions are completed.