feat: support to export data using shapefile writer
This commit adds a class to provide support to export the
source and the transformed data to shapefiles using geotools.
Also adds the plugin for the test cases for the shapefile
writer package.
Adds support to create multiple files when more than one
schema is present in the data to be exported.
Support to enable the target crs page in the shape view.
Support to still continue exporting the Shapefile export in case
there is any valid geometry in the list of files to be
exported.

ING-2971