feat: Remove single schema from project view
This commit adds support to remove single schema from the project
(Resource) view. Code is updated to store multiple schemas
in a map along with their resourceIds.
The changes are also done to fix a legacy issue to refresh the
ResourcesView when any/all schema(s) and the instances are added
or cleared.

ING-2968