fixup! feat: support to export data using shapefile writer
This commit provides a fix for exporting the Shapefiles when some
of the instances are null in the attributes which caused the
improper schema creation. The change is done to add all the
attributes even if their their data is null.
This commit also provides a fix to truncate long column names
up to 9 characters. If column names are longer than 10 characters
then the library will not export their corresponding instance data.

ING-2971