feat: add ability to skip n-lines when reading csv and xls files
This commits add the ability of skipping n-lines
when reading instances from either csv and xls files.
In the UI the user can find now a new button
to select the no. of lines to be skipped.
If one wants to skip the header,
then 1 should be selected.
If the number of lines to skip is larger
then the file length, then nothing is retrieved.
The new functionality is backwards compatible.

ING-3086