fixup! feat: enable selecting more than one file when loading schema or data
Change to simplify the autodetect checkbox and the selection box
functionality. None of these should ever be disabled. By default checkbox
should be true if multiple files are selected and false if a single file is
selected. At any point it should not happen that the checkbox is set to
true and the selection box also has data.

ING-2969