feat: add option to specify leniency (true or false) when using function DateExtraction
This commits adds the option of specifying whether or not date/time parsing is to be lenient when using the function DateExtraction. It also modifies the description of the parameter "Date format" so as to be less prone to mistakes (e.g., quotes should not be used and the format refers to the one in the source data).

ING-3214