fix: exporting instances with multiple geometry types to ShapeFiles
Code was not able to handle the scenario if the source data had
multiple geometry types because every time a new geometry type
was found, the old one was overwritten with a new object.
Code is now updated to add the schema with a new
geometry instead of overwriting the existing one.