Add option to automatically add the codespace during export. (#957)
Add an option to automatically add the codespace to the GML:identifier during export (default). The user can decide whether to use the functionality by checking the box "automatically add the codespace" during the export. The codespace is added only for feature types within INSPIRE. If the feature type is not an INSPIRE type, the field is left empty.

ING-3285