fixup: align XSL and CSV in the way of getting the selected item from the CheckboxTableViewer
Align the XSL and CSV by getting the selected elements of the CheckboxTableViewer and removing the greyed out elements that cannot be selected.
The grey elements are not used later because they don't have any properties, but better to send the complete and right list of selected elements.

related ING-3941