refactor: replace default map by OpenStreetMap
Replaces the Stamen Terrain default map by OpenStreetMap.

To comply with the [usage policy of the OpenStreetMap tile server](https://operations.osmfoundation.org/policies/tiles/),
the number of parallel connections is reduced to two and a HTTP User-Agent
header is added to the requests.

The old default map configuration, if present, is automatically removed and
the new default configuration is added.

ING-4081