feat: support reading multiple Excel sheets at once
A new setting was introduced that enables loading multiple sheets when
reading the Excel file. Type names are tried to be determined based on
match with the sheet name, if no other information is provided.

An additional new setting for the Excel reader was introduced to be able
to control in detail how an individual sheet should be loaded.

Currently support is only provided in headless mode. In the hale studio
UI it is not yet supported to load multiple sheets.

Also includes fixes:

- XLSInstanceWriter did not state correctly if it requires multiple
  iterations over the instance collection
- reading .xlsx files with hale studio, which did not work before
  because always .xls format was assumed
- test for writing multiple sheets
- provider parameter definitions for reader and writer being mixed up

ING-4062