fix: Update decimal places to 20 for Geotools
For hale studio it does not make sense to limit
decimal places so that the original data remains
intact. However, we need to provide some value
to Geotools and it does not make sense to configure
max integer. Therefore, providing higher value of 20
to Geotools even if it represents some kind of ridiculous 
"accuracy" for WGS 84.