feat: add formatter for the date
hale: add new Excel reader parameter that allows to specify import format for date cells
At the moment the formatter has the following format: yyyy-MM-dd, but should be customizable by the user with a new issue.

ING-4151

feat: add formatter for the date

hale: add new Excel reader parameter that allows to specify import format for date cells
At the moment the formatter has the following format: yyyy-MM-dd, but should be customizable by the user with a new issue.

ING-4151