fix: fix provider definition for geometry.unifyWindingOrder parameters
Up to now the `providerParameter` definitions for the parameter
`geometry.unifyWindingOrder` contained values that were inconsistent
with the corresponding enumeration `EnumWindingOrderTypes`.

Using the inconsistent values in CLI transformations could lead to unexpected
results when using the value of the parameters given in the documentation that
is derived from the `providerParameter` definitions.